* libtiff/tif_win32.c: Eliminate some minor 64-bit warnings in tif_win32.c.  Patch by Edward Lam.

+++ libtiff/tif_win32.c
@@ -377,7 +377,7 @@ Win32WarningHandler(const char* module, const char* fmt, va_list ap)
 	LPCTSTR szTitleText = "%s Warning";
 	LPCTSTR szDefaultModule = "LIBTIFF";
 	LPCTSTR szTmpModule = (module == NULL) ? szDefaultModule : module;
-        int nBufSize = (strlen(szTmpModule) +
+        SIZE_T nBufSize = (strlen(szTmpModule) +
                         strlen(szTitleText) + strlen(fmt) + 256)*sizeof(char);
 
 	if ((szTitle = (LPTSTR)LocalAlloc(LMEM_FIXED, nBufSize)) == NULL)
@@ -408,7 +408,7 @@ Win32ErrorHandler(const char* module, const char* fmt, va_list ap)
 	LPCTSTR szTitleText = "%s Error";
 	LPCTSTR szDefaultModule = "LIBTIFF";
 	LPCTSTR szTmpModule = (module == NULL) ? szDefaultModule : module;
-        int nBufSize = (strlen(szTmpModule) +
+        SIZE_T nBufSize = (strlen(szTmpModule) +
                         strlen(szTitleText) + strlen(fmt) + 256)*sizeof(char);
 
 	if ((szTitle = (LPTSTR)LocalAlloc(LMEM_FIXED, nBufSize)) == NULL)
